@@ -1,41 +1,40 @@
 {
-	"name": "kit",
+	"name": "kit-monorepo",
 	"version": "0.0.1",
-	"description": "",
-	"main": "index.js",
+	"description": "monorepo for @sveltejs/kit and friends",
+	"private": true,
+	"type": "module",
 	"scripts": {
-		"build": "pnpm build --filter=\"@sveltejs/*\" --filter=\"create-svelte\"",
-		"test": "pnpm -r test --workspace-concurrency=1",
-		"check": "pnpm -r check",
-		"lint": "pnpm -r lint",
+		"test": "pnpm test -r --filter=./packages/* --filter=!./packages/create-svelte",
+		"test:cross-platform:dev": "pnpm run --dir packages/kit test:cross-platform:dev",
+		"test:cross-platform:build": "pnpm run --dir packages/kit test:cross-platform:build",
+		"test:vite-ecosystem-ci": "pnpm test --dir packages/kit",
+		"test:create-svelte": "pnpm run --dir packages/create-svelte test",
+		"check": "pnpm -r prepublishOnly && pnpm -r check",
+		"lint": "pnpm -r lint && eslint --cache --cache-location node_modules/.eslintcache 'packages/**/*.js'",
 		"format": "pnpm -r format",
-		"release": "pnpm publish --tag=next --filter=\"@sveltejs/*\" --filter=\"create-svelte\""
+		"precommit": "pnpm format && pnpm lint",
+		"changeset:version": "changeset version && pnpm -r generate:version && git add --all",
+		"changeset:release": "changeset publish",
+		"start": "cd sites/kit.svelte.dev && npm run dev"
 	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/sveltejs/kit.git"
-	},
-	"author": "",
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/sveltejs/kit/issues"
-	},
-	"homepage": "https://github.com/sveltejs/kit#readme",
 	"devDependencies": {
-		"@changesets/cli": "^2.14.1",
-		"@rollup/plugin-commonjs": "^17.1.0",
-		"@rollup/plugin-json": "^4.1.0",
-		"@rollup/plugin-node-resolve": "^11.2.0",
-		"@sveltejs/eslint-config": "github:sveltejs/eslint-config#v5.6.0",
-		"@typescript-eslint/eslint-plugin": "^4.19.0",
-		"@typescript-eslint/parser": "^4.19.0",
-		"eslint": "^7.22.0",
-		"eslint-plugin-import": "^2.22.1",
-		"eslint-plugin-svelte3": "^3.1.2",
-		"playwright-chromium": "^1.10.0",
-		"prettier": "2.2.1",
-		"rollup": "^2.42.3",
-		"typescript": "^4.2.3"
+		"@changesets/cli": "^2.27.1",
+		"@sveltejs/eslint-config": "^6.0.4",
+		"@svitejs/changesets-changelog-github-compact": "^1.1.0",
+		"@typescript-eslint/eslint-plugin": "^6.14.0",
+		"eslint": "^8.55.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-svelte": "^2.35.1",
+		"eslint-plugin-unicorn": "^49.0.0",
+		"playwright": "1.30.0",
+		"prettier": "^3.1.1",
+		"rollup": "^3.29.4",
+		"svelte": "^4.2.8",
+		"typescript": "^5.3.3"
 	},
-	"type": "module"
+	"packageManager": "pnpm@8.12.1",
+	"engines": {
+		"pnpm": "^8.0.0"
+	}
 }
