@@ -1,23 +1,56 @@
 {
 	"name": "@sveltejs/adapter-netlify",
-	"version": "1.0.0-next.10",
-	"main": "index.js",
+	"version": "3.0.0",
+	"description": "A SvelteKit adapter that creates a Netlify app",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/sveltejs/kit",
+		"directory": "packages/adapter-netlify"
+	},
+	"license": "MIT",
+	"homepage": "https://kit.svelte.dev",
+	"type": "module",
+	"exports": {
+		".": {
+			"types": "./index.d.ts",
+			"import": "./index.js"
+		},
+		"./package.json": "./package.json"
+	},
 	"types": "index.d.ts",
 	"files": [
 		"files",
+		"index.js",
 		"index.d.ts"
 	],
 	"scripts": {
-		"lint": "eslint --ignore-path .gitignore \"**/*.{ts,js,svelte}\" && npm run check-format",
-		"format": "prettier --write . --config ../../.prettierrc --ignore-path .gitignore",
-		"check-format": "prettier --check . --config ../../.prettierrc --ignore-path .gitignore"
+		"dev": "node -e \"fs.rmSync('files', { force: true, recursive: true })\" && rollup -cw",
+		"build": "node -e \"fs.rmSync('files', { force: true, recursive: true })\" && rollup -c && cp src/edge.js files/edge.js",
+		"test": "vitest run",
+		"check": "tsc",
+		"lint": "prettier --check .",
+		"format": "pnpm lint --write",
+		"prepublishOnly": "pnpm build"
 	},
 	"dependencies": {
-		"@sveltejs/kit": "workspace:*",
-		"esbuild": "^0.11.12",
-		"toml": "^3.0.0"
+		"@iarna/toml": "^2.2.5",
+		"esbuild": "^0.19.9",
+		"set-cookie-parser": "^2.6.0"
 	},
 	"devDependencies": {
-		"typescript": "^4.2.3"
+		"@netlify/functions": "^2.4.1",
+		"@rollup/plugin-commonjs": "^25.0.7",
+		"@rollup/plugin-json": "^6.1.0",
+		"@rollup/plugin-node-resolve": "^15.2.3",
+		"@sveltejs/kit": "workspace:^",
+		"@sveltejs/vite-plugin-svelte": "^3.0.1",
+		"@types/node": "^18.19.3",
+		"@types/set-cookie-parser": "^2.4.7",
+		"rollup": "^4.8.0",
+		"typescript": "^5.3.3",
+		"vitest": "^1.0.4"
+	},
+	"peerDependencies": {
+		"@sveltejs/kit": "^2.0.0"
 	}
 }
