@@ -1,24 +1,45 @@
 {
 	"name": "@sveltejs/adapter-static",
-	"version": "1.0.0-next.7",
+	"version": "3.0.0",
+	"description": "Adapter for SvelteKit apps that prerenders your entire site as a collection of static files",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/sveltejs/kit",
+		"directory": "packages/adapter-static"
+	},
+	"license": "MIT",
+	"homepage": "https://kit.svelte.dev",
+	"type": "module",
+	"exports": {
+		".": {
+			"types": "./index.d.ts",
+			"import": "./index.js"
+		},
+		"./package.json": "./package.json"
+	},
 	"types": "index.d.ts",
+	"files": [
+		"index.js",
+		"index.d.ts",
+		"platforms.js"
+	],
 	"scripts": {
-		"lint": "eslint --ignore-path .gitignore \"**/*.{ts,js,svelte}\" && npm run check-format",
+		"lint": "prettier --check .",
 		"check": "tsc",
-		"format": "prettier --write . --config ../../.prettierrc --ignore-path .gitignore",
-		"check-format": "prettier --check . --config ../../.prettierrc --ignore-path .gitignore",
-		"test": "uvu test test.js"
+		"format": "pnpm lint --write",
+		"test": "pnpm -r --workspace-concurrency 1 --filter=\"./test/**\" test"
 	},
 	"devDependencies": {
-		"@sveltejs/kit": "workspace:*",
-		"playwright-chromium": "^1.10.0",
-		"port-authority": "^1.1.2",
-		"sirv": "^1.0.11",
-		"typescript": "^4.2.4"
+		"@playwright/test": "1.30.0",
+		"@sveltejs/kit": "workspace:^",
+		"@sveltejs/vite-plugin-svelte": "^3.0.1",
+		"@types/node": "^18.19.3",
+		"sirv": "^2.0.3",
+		"svelte": "^4.2.8",
+		"typescript": "^5.3.3",
+		"vite": "^5.0.8"
 	},
-	"type": "module",
-	"main": "index.cjs",
-	"exports": {
-		"require": "./index.cjs"
+	"peerDependencies": {
+		"@sveltejs/kit": "^2.0.0"
 	}
 }
