@@ -1,16 +1,20 @@
 {
 	"name": "~TODO~",
 	"version": "0.0.1",
+	"private": true,
 	"scripts": {
-		"dev": "svelte-kit dev",
-		"build": "svelte-kit build",
-		"start": "svelte-kit start"
+		"dev": "vite dev",
+		"build": "vite build --mode staging",
+		"preview": "sirv -p 5173 -s 200.html build",
+		"test": "playwright test"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-node": "next",
-		"@sveltejs/kit": "next",
-		"svelte": "^3.29.0",
-		"vite": "^2.1.0"
+		"@sveltejs/adapter-node": "workspace:^",
+		"@sveltejs/kit": "workspace:^",
+		"@sveltejs/vite-plugin-svelte": "^3.0.1",
+		"sirv-cli": "^2.0.2",
+		"svelte": "^4.2.8",
+		"vite": "^5.0.8"
 	},
 	"type": "module"
 }
