@@ -1,6 +1,6 @@
 {
 	"devDependencies": {
-		"eslint": "^7.22.0",
-		"eslint-plugin-svelte3": "^3.2.0"
+		"eslint": "^8.28.0",
+		"eslint-plugin-svelte": "^2.30.0"
 	}
 }
