@@ -1,7 +1,11 @@
 {
+	"scripts": {
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
+	},
 	"devDependencies": {
-		"typescript": "^4.0.0",
-		"tslib": "^2.0.0",
-		"svelte-preprocess": "^4.0.0"
+		"typescript": "^5.0.0",
+		"tslib": "^2.4.1",
+		"svelte-check": "^3.6.0"
 	}
 }
