@@ -1,25 +1,23 @@
 {
-	"name": "~TODO~",
-	"version": "0.0.1",
+	"name": "default-template",
+	"version": "0.0.2-next.0",
+	"private": true,
 	"scripts": {
-		"dev": "svelte-kit dev",
-		"build": "svelte-kit build --verbose",
-		"start": "svelte-kit start"
+		"dev": "vite dev",
+		"build": "vite build",
+		"preview": "vite preview"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-cloudflare-workers": "next",
-		"@sveltejs/adapter-netlify": "next",
-		"@sveltejs/adapter-vercel": "next",
-		"@sveltejs/kit": "next",
-		"svelte": "^3.37.0",
-		"svelte-preprocess": "^4.7.1",
-		"typescript": "^4.2.4",
-		"vite": "^2.2.3"
+		"@neoconfetti/svelte": "^1.0.0",
+		"@sveltejs/adapter-auto": "workspace:*",
+		"@sveltejs/kit": "workspace:*",
+		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"svelte": "^4.2.8",
+		"typescript": "^5.3.3",
+		"vite": "^5.0.8"
 	},
 	"type": "module",
 	"dependencies": {
-		"@fontsource/fira-mono": "^4.2.2",
-		"@lukeed/uuid": "^2.0.0",
-		"cookie": "^0.4.1"
+		"@fontsource/fira-mono": "^5.0.8"
 	}
 }
