@@ -1,246 +1,175 @@
 #!/usr/bin/env node
-import fs from 'fs';
-import path from 'path';
-import { fileURLToPath } from 'url';
-import { bold, cyan, gray, green, red } from 'kleur/colors';
-import prompts from 'prompts';
-import { mkdirp, copy } from './utils.js';
-
-// prettier-ignore
-const disclaimer = `
-${bold(cyan('Welcome to SvelteKit!'))}
-
-${bold(red('This is beta software; expect bugs and missing features.'))}
-
-If you encounter a problem, open an issue on ${cyan('https://github.com/sveltejs/kit/issues')} if none exists already.
-`;
+import fs from 'node:fs';
+import path from 'node:path';
+import * as p from '@clack/prompts';
+import { bold, cyan, grey, yellow } from 'kleur/colors';
+import { create } from './index.js';
+import { dist, package_manager } from './utils.js';
 
 const { version } = JSON.parse(fs.readFileSync(new URL('package.json', import.meta.url), 'utf-8'));
+let cwd = process.argv[2] || '.';
 
-async function main() {
-	console.log(gray(`\ncreate-svelte version ${version}`));
-	console.log(disclaimer);
+console.log(`
+${grey(`create-svelte version ${version}`)}
+`);
 
-	const cwd = process.argv[2] || '.';
+p.intro('Welcome to SvelteKit!');
 
-	if (fs.existsSync(cwd)) {
-		if (fs.readdirSync(cwd).length > 0) {
-			const response = await prompts({
-				type: 'confirm',
-				name: 'value',
-				message: 'Directory not empty. Continue?',
-				initial: false
-			});
-
-			if (!response.value) {
-				process.exit(1);
-			}
-		}
-	} else {
-		mkdirp(cwd);
-	}
-
-	const options = /** @type {import('./types/internal').Options} */ (await prompts([
-		{
-			type: 'select',
-			name: 'template',
-			message: 'Which Svelte app template?',
-			choices: fs.readdirSync(dist('templates')).map((dir) => {
-				const meta_file = dist(`templates/${dir}/meta.json`);
-				const meta = JSON.parse(fs.readFileSync(meta_file, 'utf8'));
-
-				return {
-					title: meta.description,
-					value: dir
-				};
-			})
-		},
-		{
-			type: 'confirm',
-			name: 'typescript',
-			message: 'Use TypeScript?',
-			initial: false
-		},
-		{
-			type: 'confirm',
-			name: 'eslint',
-			message: 'Add ESLint for code linting?',
-			initial: false
-		},
-		{
-			type: 'confirm',
-			name: 'prettier',
-			message: 'Add Prettier for code formatting?',
-			initial: false
-		}
-	]));
-
-	const name = path.basename(path.resolve(cwd));
-
-	write_template_files(options.template, options.typescript, name, cwd);
-	write_common_files(cwd, options);
-
-	console.log(bold(green('✔ Copied project files')));
-
-	if (options.typescript) {
-		console.log(
-			bold(
-				green(
-					'✔ Added TypeScript support. ' +
-						'To use it inside Svelte components, add lang="ts" to the attributes of a script tag.'
-				)
-			)
-		);
-	}
-
-	if (options.eslint) {
-		console.log(
-			bold(
-				green(
-					'✔ Added ESLint.\n' +
-						'Readme for ESLint and Svelte: https://github.com/sveltejs/eslint-plugin-svelte3'
-				)
-			)
-		);
-	}
-
-	if (options.prettier) {
-		console.log(
-			bold(
-				green(
-					'✔ Added Prettier.\n' +
-						'General formatting options: https://prettier.io/docs/en/options.html\n' +
-						'Svelte-specific formatting options: https://github.com/sveltejs/prettier-plugin-svelte#options'
-				)
-			)
-		);
-	}
-
-	console.log(
-		'\nWant to add other parts to your code base? ' +
-			'Visit https://github.com/svelte-add/svelte-adders, a community project of commands ' +
-			'to add particular functionality to Svelte projects\n'
-	);
-
-	console.log('\nNext steps:');
-	let i = 1;
-
-	const relative = path.relative(process.cwd(), cwd);
-	if (relative !== '') {
-		console.log(`  ${i++}: ${bold(cyan(`cd ${relative}`))}`);
-	}
-
-	// prettier-ignore
-	console.log(`  ${i++}: ${bold(cyan('git init && git add -A && git commit -m "Initial commit"'))} (optional step)`);
-	console.log(`  ${i++}: ${bold(cyan('npm install'))} (or pnpm install, etc)`);
-	console.log(`  ${i++}: ${bold(cyan('npm run dev -- --open'))}`);
-
-	console.log(`\nTo close the dev server, hit ${bold(cyan('Ctrl-C'))}`);
-	console.log('\nStuck? Visit us at https://svelte.dev/chat\n');
-}
-
-/**
- * @param {string} template
- * @param {boolean} typescript
- * @param {string} name
- * @param {string} cwd
- */
-function write_template_files(template, typescript, name, cwd) {
-	const dir = dist(`templates/${template}`);
-	copy(`${dir}/assets`, cwd, (name) => name.replace('gitignore', '.gitignore'));
-	copy(`${dir}/package.json`, `${cwd}/package.json`);
-
-	const manifest = `${dir}/files.${typescript ? 'ts' : 'js'}.json`;
-	const files = /** @type {import('./types/internal').File[]} */ (JSON.parse(
-		fs.readFileSync(manifest, 'utf-8')
-	));
-
-	files.forEach((file) => {
-		const dest = path.join(cwd, file.name);
-		mkdirp(path.dirname(dest));
-
-		fs.writeFileSync(dest, file.contents.replace(/~TODO~/g, name));
-	});
-}
-
-/**
- *
- * @param {string} cwd
- * @param {import('./types/internal').Options} options
- */
-function write_common_files(cwd, options) {
-	const shared = dist('shared.json');
-	const { files } = /** @type {import('./types/internal').Common} */ (JSON.parse(
-		fs.readFileSync(shared, 'utf-8')
-	));
-
-	const pkg_file = path.join(cwd, 'package.json');
-	const pkg = /** @type {any} */ (JSON.parse(fs.readFileSync(pkg_file, 'utf-8')));
-
-	files.forEach((file) => {
-		const include = file.include.every((condition) => options[condition]);
-		const exclude = file.exclude.some((condition) => options[condition]);
-
-		if (exclude || !include) return;
-
-		if (file.name === 'package.json') {
-			const new_pkg = JSON.parse(file.contents);
-			merge(pkg, new_pkg);
-		} else {
-			const dest = path.join(cwd, file.name);
-			mkdirp(path.dirname(dest));
-			fs.writeFileSync(dest, file.contents);
-		}
+if (cwd === '.') {
+	const dir = await p.text({
+		message: 'Where should we create your project?',
+		placeholder: '  (hit Enter to use current directory)'
 	});
 
-	pkg.dependencies = sort_keys(pkg.dependencies);
-	pkg.devDependencies = sort_keys(pkg.devDependencies);
+	if (p.isCancel(dir)) process.exit(1);
 
-	fs.writeFileSync(pkg_file, JSON.stringify(pkg, null, '  '));
-}
-
-/**
- * @param {any} target
- * @param {any} source
- */
-function merge(target, source) {
-	for (const key in source) {
-		if (key in target) {
-			const target_value = target[key];
-			const source_value = source[key];
-
-			if (
-				typeof source_value !== typeof target_value ||
-				Array.isArray(source_value) !== Array.isArray(target_value)
-			) {
-				throw new Error('Mismatched values');
-			}
-
-			merge(target_value, source_value);
-		} else {
-			target[key] = source[key];
-		}
+	if (dir) {
+		cwd = /** @type {string} */ (dir);
 	}
 }
 
-/** @param {Record<string, any>} obj */
-function sort_keys(obj) {
-	if (!obj) return;
-
-	/** @type {Record<string, any>} */
-	const sorted = {};
-	Object.keys(obj)
-		.sort()
-		.forEach((key) => {
-			sorted[key] = obj[key];
+if (fs.existsSync(cwd)) {
+	if (fs.readdirSync(cwd).length > 0) {
+		const force = await p.confirm({
+			message: 'Directory not empty. Continue?',
+			initialValue: false
 		});
 
-	return sorted;
+		// bail if `force` is `false` or the user cancelled with Ctrl-C
+		if (force !== true) {
+			process.exit(1);
+		}
+	}
 }
 
-/** @param {string} path */
-function dist(path) {
-	return fileURLToPath(new URL(`./dist/${path}`, import.meta.url).href);
+const options = await p.group(
+	{
+		template: () =>
+			p.select({
+				message: 'Which Svelte app template?',
+				// @ts-expect-error i have no idea what is going on here
+				options: fs.readdirSync(dist('templates')).map((dir) => {
+					const meta_file = dist(`templates/${dir}/meta.json`);
+					const { title, description } = JSON.parse(fs.readFileSync(meta_file, 'utf8'));
+
+					return {
+						label: title,
+						hint: description,
+						value: dir
+					};
+				})
+			}),
+
+		types: () =>
+			p.select({
+				message: 'Add type checking with TypeScript?',
+				initialValue: /** @type {'checkjs' | 'typescript' | null} */ ('checkjs'),
+				options: [
+					{
+						label: 'Yes, using JavaScript with JSDoc comments',
+						value: 'checkjs'
+					},
+					{
+						label: 'Yes, using TypeScript syntax',
+						value: 'typescript'
+					},
+					{ label: 'No', value: null }
+				]
+			}),
+
+		features: () =>
+			p.multiselect({
+				message: 'Select additional options (use arrow keys/space bar)',
+				required: false,
+				options: [
+					{
+						value: 'eslint',
+						label: 'Add ESLint for code linting'
+					},
+					{
+						value: 'prettier',
+						label: 'Add Prettier for code formatting'
+					},
+					{
+						value: 'playwright',
+						label: 'Add Playwright for browser testing'
+					},
+					{
+						value: 'vitest',
+						label: 'Add Vitest for unit testing'
+					},
+					{
+						value: 'svelte5',
+						label: 'Try the Svelte 5 preview (unstable!)'
+					}
+				]
+			})
+	},
+	{ onCancel: () => process.exit(1) }
+);
+
+await create(cwd, {
+	name: path.basename(path.resolve(cwd)),
+	template: /** @type {'default' | 'skeleton' | 'skeletonlib'} */ (options.template),
+	types: options.types,
+	prettier: options.features.includes('prettier'),
+	eslint: options.features.includes('eslint'),
+	playwright: options.features.includes('playwright'),
+	vitest: options.features.includes('vitest'),
+	svelte5: options.features.includes('svelte5')
+});
+
+p.outro('Your project is ready!');
+
+if (options.types === 'typescript') {
+	console.log(bold('✔ Typescript'));
+	console.log('  Inside Svelte components, use <script lang="ts">\n');
+} else if (options.types === 'checkjs') {
+	console.log(bold('✔ Type-checked JavaScript'));
+	console.log(cyan('  https://www.typescriptlang.org/tsconfig#checkJs\n'));
+} else if (options.template === 'skeletonlib') {
+	const warning = yellow('▲');
+	console.log(
+		`${warning} You chose to not add type checking, but TypeScript will still be installed in order to generate type definitions when building the library\n`
+	);
 }
 
-main();
+if (options.features.includes('eslint')) {
+	console.log(bold('✔ ESLint'));
+	console.log(cyan('  https://github.com/sveltejs/eslint-plugin-svelte\n'));
+}
+
+if (options.features.includes('prettier')) {
+	console.log(bold('✔ Prettier'));
+	console.log(cyan('  https://prettier.io/docs/en/options.html'));
+	console.log(cyan('  https://github.com/sveltejs/prettier-plugin-svelte#options\n'));
+}
+
+if (options.features.includes('playwright')) {
+	console.log(bold('✔ Playwright'));
+	console.log(cyan('  https://playwright.dev\n'));
+}
+
+if (options.features.includes('vitest')) {
+	console.log(bold('✔ Vitest'));
+	console.log(cyan('  https://vitest.dev\n'));
+}
+
+console.log('Install community-maintained integrations:');
+console.log(cyan('  https://github.com/svelte-add/svelte-add'));
+
+console.log('\nNext steps:');
+let i = 1;
+
+const relative = path.relative(process.cwd(), cwd);
+if (relative !== '') {
+	console.log(`  ${i++}: ${bold(cyan(`cd ${relative}`))}`);
+}
+
+console.log(`  ${i++}: ${bold(cyan(`${package_manager} install`))}`);
+// prettier-ignore
+console.log(`  ${i++}: ${bold(cyan('git init && git add -A && git commit -m "Initial commit"'))} (optional)`);
+console.log(`  ${i++}: ${bold(cyan(`${package_manager} run dev -- --open`))}`);
+
+console.log(`\nTo close the dev server, hit ${bold(cyan('Ctrl-C'))}`);
+console.log(`\nStuck? Visit us at ${cyan('https://svelte.dev/chat')}`);
