@@ -1,8 +1,12 @@
 export type Options = {
-	template: 'default' | 'skeleton';
-	typescript: boolean;
+	name: string;
+	template: 'default' | 'skeleton' | 'skeletonlib';
+	types: 'typescript' | 'checkjs' | null;
 	prettier: boolean;
 	eslint: boolean;
+	playwright: boolean;
+	vitest: boolean;
+	svelte5?: boolean; // optional to not introduce a breaking change to the `create` API
 };
 
 export type File = {
@@ -10,11 +14,23 @@ export type File = {
 	contents: string;
 };
 
+export type Condition =
+	| 'eslint'
+	| 'prettier'
+	| 'typescript'
+	| 'checkjs'
+	| 'playwright'
+	| 'vitest'
+	| 'skeleton'
+	| 'default'
+	| 'skeletonlib'
+	| 'svelte5';
+
 export type Common = {
 	files: Array<{
 		name: string;
-		include: Array<'eslint' | 'prettier' | 'typescript'>;
-		exclude: Array<'eslint' | 'prettier' | 'typescript'>;
+		include: Condition[];
+		exclude: Condition[];
 		contents: string;
 	}>;
 };
