@@ -1,12 +1,13 @@
-import fs from 'fs';
-import path from 'path';
+import fs from 'node:fs';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
 
 /** @param {string} dir */
 export function mkdirp(dir) {
 	try {
 		fs.mkdirSync(dir, { recursive: true });
 	} catch (e) {
-		if (e.code === 'EEXIST') return;
+		if (/** @type {any} */ (e).code === 'EEXIST') return;
 		throw e;
 	}
 }
@@ -42,4 +43,28 @@ export function copy(from, to, rename = identity) {
 		mkdirp(path.dirname(to));
 		fs.copyFileSync(from, to);
 	}
+}
+
+/** @param {string} path */
+export function dist(path) {
+	return fileURLToPath(new URL(`./dist/${path}`, import.meta.url).href);
+}
+
+/** @type {string} */
+export const package_manager = get_package_manager() || 'npm';
+
+/**
+ * Supports npm, pnpm, Yarn, cnpm, bun and any other package manager that sets the
+ * npm_config_user_agent env variable.
+ * Thanks to https://github.com/zkochan/packages/tree/main/which-pm-runs for this code!
+ */
+function get_package_manager() {
+	if (!process.env.npm_config_user_agent) {
+		return undefined;
+	}
+	const user_agent = process.env.npm_config_user_agent;
+	const pm_spec = user_agent.split(' ')[0];
+	const separator_pos = pm_spec.lastIndexOf('/');
+	const name = pm_spec.substring(0, separator_pos);
+	return name === 'npminstall' ? 'cnpm' : name;
 }
