@@ -1,38 +1,40 @@
-import fs from 'fs';
-import path from 'path';
+import fs from 'node:fs';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
 import colors from 'kleur';
-import { copy } from './filesystem/index.js';
-import { fileURLToPath } from 'url';
+import { posixify, to_fs } from '../utils/filesystem.js';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+/**
+ * Resolved path of the `runtime` directory
+ *
+ * TODO Windows issue:
+ * Vite or sth else somehow sets the driver letter inconsistently to lower or upper case depending on the run environment.
+ * In playwright debug mode run through VS Code this a root-to-lowercase conversion is needed in order for the tests to run.
+ * If we do this conversion in other cases it has the opposite effect though and fails.
+ */
+export const runtime_directory = posixify(fileURLToPath(new URL('../runtime', import.meta.url)));
 
-/** @param {string} dest */
-export function copy_assets(dest) {
-	let prefix = '..';
-	do {
-		// we jump through these hoops so that this function
-		// works whether or not it's been bundled
-		const resolved = path.resolve(__dirname, `${prefix}/assets`);
-
-		if (fs.existsSync(resolved)) {
-			copy(resolved, dest);
-			return;
-		}
-
-		prefix = `../${prefix}`;
-	} while (true); // eslint-disable-line
-}
+/**
+ * This allows us to import SvelteKit internals that aren't exposed via `pkg.exports` in a
+ * way that works whether `@sveltejs/kit` is installed inside the project's `node_modules`
+ * or in a workspace root
+ */
+export const runtime_base = runtime_directory.startsWith(process.cwd())
+	? `/${path.relative('.', runtime_directory)}`
+	: to_fs(runtime_directory);
 
 function noop() {}
 
 /** @param {{ verbose: boolean }} opts */
 export function logger({ verbose }) {
-	/** @type {import('types/internal').Logger} */
+	/** @type {import('types').Logger} */
 	const log = (msg) => console.log(msg.replace(/^/gm, '  '));
 
+	/** @param {string} msg */
+	const err = (msg) => console.error(msg.replace(/^/gm, '  '));
+
 	log.success = (msg) => log(colors.green(`✔ ${msg}`));
-	log.error = (msg) => log(colors.bold().red(msg));
+	log.error = (msg) => err(colors.bold().red(msg));
 	log.warn = (msg) => log(colors.bold().yellow(msg));
 
 	log.minor = verbose ? (msg) => log(colors.grey(msg)) : noop;
@@ -41,68 +43,44 @@ export function logger({ verbose }) {
 	return log;
 }
 
-/**
- * Given an entry point like [cwd]/src/hooks, returns a filename like [cwd]/src/hooks.js or [cwd]/src/hooks/index.js
- * @param {string} entry
- * @returns {string}
- */
-export function resolve_entry(entry) {
-	if (fs.existsSync(entry)) {
-		const stats = fs.statSync(entry);
-		if (stats.isDirectory()) {
-			return resolve_entry(path.join(entry, 'index'));
+/** @param {import('types').ManifestData} manifest_data */
+export function get_mime_lookup(manifest_data) {
+	/** @type {Record<string, string>} */
+	const mime = {};
+
+	manifest_data.assets.forEach((asset) => {
+		if (asset.type) {
+			const ext = path.extname(asset.file);
+			mime[ext] = asset.type;
 		}
+	});
 
-		return entry;
-	} else {
-		const dir = path.dirname(entry);
+	return mime;
+}
 
-		if (fs.existsSync(dir)) {
-			const base = path.basename(entry);
-			const files = fs.readdirSync(dir);
+/**
+ * @param {string} dir
+ * @param {(file: string) => boolean} [filter]
+ */
+export function list_files(dir, filter) {
+	/** @type {string[]} */
+	const files = [];
 
-			const found = files.find((file) => file.replace(/\.[^.]+$/, '') === base);
-
-			if (found) return path.join(dir, found);
+	/** @param {string} current */
+	function walk(current) {
+		for (const file of fs.readdirSync(path.resolve(dir, current))) {
+			const child = path.posix.join(current, file);
+			if (fs.statSync(path.resolve(dir, child)).isDirectory()) {
+				walk(child);
+			} else {
+				if (!filter || filter(child)) {
+					files.push(child);
+				}
+			}
 		}
 	}
 
-	return null;
-}
+	if (fs.existsSync(dir)) walk('');
 
-/** @param {string} str */
-export function posixify(str) {
-	return str.replace(/\\/g, '/');
-}
-
-/**
- * Get a list of packages that use pkg.svelte, so they can be added
- * to ssr.noExternal. This is done on a best-effort basis to reduce
- * the frequency of 'Must use import to load ES Module' and similar
- * @param {string} cwd
- * @returns {string[]}
- */
-function find_svelte_packages(cwd) {
-	const pkg_file = path.join(cwd, 'package.json');
-	if (!fs.existsSync(pkg_file)) return [];
-
-	const pkg = JSON.parse(fs.readFileSync(pkg_file, 'utf8'));
-
-	const deps = [...Object.keys(pkg.dependencies || {}), ...Object.keys(pkg.devDependencies || {})];
-
-	return deps.filter((dep) => {
-		const dep_pkg_file = path.join(cwd, 'node_modules', dep, 'package.json');
-		if (!fs.existsSync(dep_pkg_file)) return false;
-
-		const dep_pkg = JSON.parse(fs.readFileSync(dep_pkg_file, 'utf-8'));
-		return !!dep_pkg.svelte;
-	});
-}
-
-/**
- * @param {string} cwd
- * @param {string[]} [user_specified_deps]
- */
-export function get_no_external(cwd, user_specified_deps = []) {
-	return [...user_specified_deps, ...find_svelte_packages(cwd)];
+	return files;
 }
