@@ -1,14 +1,1 @@
-/**
- * @type {import('$app/env').browser}
- */
-export const browser = !import.meta.env.SSR;
-/**
- * @type {import('$app/env').dev}
- */
-export const dev = !!import.meta.env.DEV;
-/**
- * @type {import('$app/env').amp}
- */
-export const amp = !!import.meta.env.VITE_SVELTEKIT_AMP;
-
-export { prerendering } from '../env.js';
+throw new Error('$app/env has been renamed to $app/environment');
