@@ -1,85 +1,94 @@
 import { getContext } from 'svelte';
-
-// const ssr = (import.meta as any).env.SSR;
-const ssr = typeof window === 'undefined'; // TODO why doesn't previous line work in build?
-
-// TODO remove this (for 1.0? after 1.0?)
-let warned = false;
-export function stores() {
-	if (!warned) {
-		console.error('stores() is deprecated; use getStores() instead');
-		warned = true;
-	}
-	return getStores();
-}
+import { browser } from './environment.js';
+import { stores as browser_stores } from '../client/singletons.js';
 
 /**
- * @type {import('$app/stores').getStores}
+ * A function that returns all of the contextual stores. On the server, this must be called during component initialization.
+ * Only use this if you need to defer store subscription until after the component has mounted, for some reason.
  */
 export const getStores = () => {
-	const stores = getContext('__svelte__');
+	const stores = browser ? browser_stores : getContext('__svelte__');
 
 	return {
+		/** @type {typeof page} */
 		page: {
 			subscribe: stores.page.subscribe
 		},
+		/** @type {typeof navigating} */
 		navigating: {
 			subscribe: stores.navigating.subscribe
 		},
-		// @ts-ignore - deprecated, not part of type definitions, but still callable
-		get preloading() {
-			console.error('stores.preloading is deprecated; use stores.navigating instead');
-			return {
-				subscribe: stores.navigating.subscribe
-			};
-		},
-		session: stores.session
+		/** @type {typeof updated} */
+		updated: stores.updated
 	};
 };
 
-/** @type {typeof import('$app/stores').page} */
+/**
+ * A readable store whose value contains page data.
+ *
+ * On the server, this store can only be subscribed to during component initialization. In the browser, it can be subscribed to at any time.
+ *
+ * @type {import('svelte/store').Readable<import('@sveltejs/kit').Page>}
+ */
 export const page = {
-	/** @param {(value: any) => void} fn */
 	subscribe(fn) {
-		const store = getStores().page;
+		const store = __SVELTEKIT_DEV__ ? get_store('page') : getStores().page;
 		return store.subscribe(fn);
 	}
 };
 
-/** @type {typeof import('$app/stores').navigating} */
+/**
+ * A readable store.
+ * When navigating starts, its value is a `Navigation` object with `from`, `to`, `type` and (if `type === 'popstate'`) `delta` properties.
+ * When navigating finishes, its value reverts to `null`.
+ *
+ * On the server, this store can only be subscribed to during component initialization. In the browser, it can be subscribed to at any time.
+ * @type {import('svelte/store').Readable<import('@sveltejs/kit').Navigation | null>}
+ */
 export const navigating = {
-	/** @param {(value: any) => void} fn */
 	subscribe(fn) {
-		const store = getStores().navigating;
+		const store = __SVELTEKIT_DEV__ ? get_store('navigating') : getStores().navigating;
 		return store.subscribe(fn);
 	}
 };
 
-/** @param {string} verb */
-const error = (verb) => {
-	throw new Error(
-		ssr
-			? `Can only ${verb} session store in browser`
-			: `Cannot ${verb} session store before subscribing`
-	);
-};
-
-/** @type {typeof import('$app/stores').session} */
-export const session = {
+/**
+ * A readable store whose initial value is `false`. If [`version.pollInterval`](https://kit.svelte.dev/docs/configuration#version) is a non-zero value, SvelteKit will poll for new versions of the app and update the store value to `true` when it detects one. `updated.check()` will force an immediate check, regardless of polling.
+ *
+ * On the server, this store can only be subscribed to during component initialization. In the browser, it can be subscribed to at any time.
+ * @type {import('svelte/store').Readable<boolean> & { check(): Promise<boolean> }}
+ */
+export const updated = {
 	subscribe(fn) {
-		const store = getStores().session;
+		const store = __SVELTEKIT_DEV__ ? get_store('updated') : getStores().updated;
 
-		if (!ssr) {
-			session.set = store.set;
-			session.update = store.update;
+		if (browser) {
+			updated.check = store.check;
 		}
 
 		return store.subscribe(fn);
 	},
-	set: (value) => {
-		error('set');
-	},
-	update: (updater) => {
-		error('update');
+	check: () => {
+		throw new Error(
+			browser
+				? 'Cannot check updated store before subscribing'
+				: 'Can only check updated store in browser'
+		);
 	}
 };
+
+/**
+ * @template {keyof ReturnType<typeof getStores>} Name
+ * @param {Name} name
+ * @returns {ReturnType<typeof getStores>[Name]}
+ */
+function get_store(name) {
+	try {
+		return getStores()[name];
+	} catch (e) {
+		throw new Error(
+			`Cannot subscribe to '${name}' store on the server outside of a Svelte component, as it is bound to the current request via component context. This prevents state from leaking between users.` +
+				'For more information, see https://kit.svelte.dev/docs/state-management#avoid-shared-state-on-the-server'
+		);
+	}
+}
