@@ -1,58 +1,28 @@
-// @ts-ignore
-import Root from 'ROOT'; // eslint-disable-line import/no-unresolved
-// @ts-ignore
-import { routes, fallback } from 'MANIFEST'; // eslint-disable-line import/no-unresolved
-import { Router } from './router.js';
-import { Renderer } from './renderer.js';
+import { DEV } from 'esm-env';
+import { create_client } from './client.js';
 import { init } from './singletons.js';
-import { set_paths } from '../paths.js';
 
-/** @param {{
- *   paths: {
- *     assets: string;
- *     base: string;
- *   },
- *   target: Node;
- *   session: any;
- *   host: string;
- *   route: boolean;
- *   spa: boolean;
- *   hydrate: {
- *     status: number;
- *     error: Error;
- *     nodes: Array<Promise<import('types/internal').CSRComponent>>;
- *     page: import('types/page').Page;
- *   };
- * }} opts */
-export async function start({ paths, target, session, host, route, spa, hydrate }) {
-	const router =
-		route &&
-		new Router({
-			base: paths.base,
-			routes
-		});
-
-	const renderer = new Renderer({
-		Root,
-		fallback,
-		target,
-		session,
-		host
-	});
-
-	init(router);
-	set_paths(paths);
-
-	if (hydrate) await renderer.start(hydrate);
-	if (route) router.init(renderer);
-
-	if (spa) router.goto(location.href, { replaceState: true }, []);
-
-	dispatchEvent(new CustomEvent('sveltekit:start'));
-}
-
-if (import.meta.env.VITE_SVELTEKIT_SERVICE_WORKER) {
-	if ('serviceWorker' in navigator) {
-		navigator.serviceWorker.register(import.meta.env.VITE_SVELTEKIT_SERVICE_WORKER);
+/**
+ * @param {import('./types.js').SvelteKitApp} app
+ * @param {HTMLElement} target
+ * @param {Parameters<import('./types.js').Client['_hydrate']>[0]} [hydrate]
+ */
+export async function start(app, target, hydrate) {
+	if (DEV && target === document.body) {
+		console.warn(
+			'Placing %sveltekit.body% directly inside <body> is not recommended, as your app may break for users who have certain browser extensions installed.\n\nConsider wrapping it in an element:\n\n<div style="display: contents">\n  %sveltekit.body%\n</div>'
+		);
 	}
+
+	const client = create_client(app, target);
+
+	init({ client });
+
+	if (hydrate) {
+		await client._hydrate(hydrate);
+	} else {
+		client.goto(location.href, { replaceState: true });
+	}
+
+	client._start_router();
 }
