@@ -12,8 +12,11 @@ declare global {
 		}
 	) => Promise<void>;
 
-	const prefetch: (url: string) => Promise<void>;
-	const prefetchRoutes: (urls?: string[]) => Promise<void>;
+	const invalidate: (url: string) => Promise<void>;
+	const preloadData: (url: string) => Promise<void>;
+	const beforeNavigate: (fn: (url: URL) => void | boolean) => void;
+	const afterNavigate: (fn: () => void) => void;
+	const preloadCode: (...urls: string[]) => Promise<void>;
 }
 
 export {};
